# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Digital Marketing Command Center API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List all brand properties
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create a brand property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyBody"
      responses:
        "201":
          description: Created property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get a property by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
    patch:
      operationId: updateProperty
      tags: [properties]
      summary: Update a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePropertyBody"
      responses:
        "200":
          description: Updated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # Agents
...
  /properties/{id}/analytics-data:
    get:
      operationId: getPropertyAnalyticsData
      tags: [analytics]
      summary: Fetch live analytics data from all connected platforms for a property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: dateRange
          in: query
          required: false
          schema:
            type: string
            enum: [7days, 30days, 90days]
            default: 30days
      responses:
        "200":
          description: Aggregated analytics data from all connected platforms
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnalyticsData"
        "404":
          description: Property not found

  # Schedules
...
components:
  schemas:
...
    Property:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        brandVoice:
          type: ["string", "null"]
        tone:
          type: ["string", "null"]
        targetAudience:
          type: ["string", "null"]
        primaryKeywords:
          type: ["string", "null"]
        websiteUrl:
          type: ["string", "null"]
        instagramHandle:
          type: ["string", "null"]
        facebookHandle:
          type: ["string", "null"]
        twitterHandle:
          type: ["string", "null"]
        linkedinHandle:
          type: ["string", "null"]
        googleAdsCustomerId:
          type: ["string", "null"]
          description: "Google Ads customer/account ID (e.g. 123-456-7890) — read-only safe field"
        metaAdsAccountId:
          type: ["string", "null"]
          description: "Meta Ads account ID (e.g. act_123456789) — read-only safe field"
        metaAdPageId:
          type: ["string", "null"]
          description: "Meta (Facebook) Business Page ID used for ad creative creation"
        wordpressConfigured:
          type: boolean
        squarespaceConfigured:
          type: boolean
        googleAdsConfigured:
          type: boolean
          description: "True when both Google Ads customer ID and refresh token are stored"
        metaAdsConfigured:
          type: boolean
          description: "True when both Meta Ads account ID and access token are stored"
        openedAt:
          type: ["string", "null"]
          description: "Date the property opened (ISO 8601)"
        propertyType:
          type: ["string", "null"]
          description: "Category of the property, e.g. Restaurant, Wine Shop, Bar"
        logoUrl:
          type: ["string", "null"]
          description: "Base64 data URL or hosted URL for the property logo"
        resyUrl:
          type: ["string", "null"]
          description: "Resy reservation page URL for this property"
        hubspotPortalId:
          type: ["string", "null"]
          description: "HubSpot portal/account ID"
        hubspotConfigured:
          type: boolean
          description: "True when a HubSpot API key is stored"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - wordpressConfigured
        - squarespaceConfigured
        - googleAdsConfigured
        - metaAdsConfigured
        - hubspotConfigured
        - createdAt
        - updatedAt
    CreatePropertyBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        brandVoice:
          type: string
        tone:
          type: string
        targetAudience:
          type: string
        primaryKeywords:
          type: string
        websiteUrl:
          type: string
        instagramHandle:
          type: string
        facebookHandle:
          type: string
        twitterHandle:
          type: string
        linkedinHandle:
          type: string
        wordpressUrl:
          type: string
        wordpressUsername:
          type: string
        wordpressAppPassword:
          type: string
        squarespaceApiKey:
          type: string
        squarespaceCollectionId:
          type: string
        googleAdsCustomerId:
          type: string
          description: "Google Ads customer ID (digits only, e.g. 1234567890)"
        googleAdsRefreshToken:
          type: string
          description: "Google Ads OAuth2 refresh token — write-only, never returned"
        metaAdsAccountId:
          type: string
          description: "Meta Ads account ID without 'act_' prefix, e.g. 123456789"
        metaAdsAccessToken:
          type: string
          description: "Meta Ads access token — write-only, never returned"
        metaAdPageId:
          type: string
          description: "Facebook Business Page ID required for ad creative creation"
        openedAt:
          type: string
          description: "ISO 8601 date when the property opened"
        propertyType:
          type: string
          description: "Category of the property"
        logoUrl:
          type: string
          description: "Base64 data URL for the property logo"
        resyUrl:
          type: string
          description: "Resy reservation page URL"
        hubspotPortalId:
          type: string
          description: "HubSpot portal/account ID"
        hubspotApiKey:
          type: string
          description: "HubSpot private API key — write-only, never returned"
      required:
        - name
    UpdatePropertyBody:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        brandVoice:
          type: string
        tone:
          type: string
        targetAudience:
          type: string
        primaryKeywords:
          type: string
        websiteUrl:
          type: string
        instagramHandle:
          type: string
        facebookHandle:
          type: string
        twitterHandle:
          type: string
        linkedinHandle:
          type: string
        wordpressUrl:
          type: string
        wordpressUsername:
          type: string
        wordpressAppPassword:
          type: string
        squarespaceApiKey:
          type: string
        squarespaceCollectionId:
          type: string
        googleAdsCustomerId:
          type: string
        googleAdsRefreshToken:
          type: string
          description: "Write-only — leave blank to keep existing value"
        metaAdsAccountId:
          type: string
        metaAdsAccessToken:
          type: string
          description: "Write-only — leave blank to keep existing value"
        metaAdPageId:
          type: string
          description: "Facebook Business Page ID required for ad creative creation"
        openedAt:
          type: string
          description: "ISO 8601 date when the property opened"
        propertyType:
          type: string
          description: "Category of the property"
        logoUrl:
          type: string
          description: "Base64 data URL for the property logo"
        resyUrl:
          type: string
          description: "Resy reservation page URL"
        hubspotPortalId:
          type: string
          description: "HubSpot portal/account ID"
        hubspotApiKey:
          type: string
          description: "HubSpot private API key — write-only, leave blank to keep existing value"
...
    AnalyticsData:
      type: object
      properties:
        property:
          type: object
          properties:
            id:
              type: integer
            name:
              type: string
          required:
            - id
            - name
        dateRange:
          type: object
          properties:
            from:
              type: string
            to:
              type: string
            label:
              type: string
          required:
            - from
            - to
            - label
        googleAds:
          type: object
          properties:
            available:
              type: boolean
            campaigns:
              type: array
              items:
                type: object
            totals:
              type: object
              properties:
                impressions:
                  type: integer
                clicks:
                  type: integer
                spend:
                  type: number
                ctr:
                  type: number
              required:
                - impressions
                - clicks
                - spend
                - ctr
          required:
            - available
        metaAds:
          type: object
          properties:
            available:
              type: boolean
            campaigns:
              type: array
              items:
                type: object
            totals:
              type: object
              properties:
                impressions:
                  type: integer
                clicks:
                  type: integer
                spend:
                  type: number
                ctr:
                  type: number
              required:
                - impressions
                - clicks
                - spend
                - ctr
          required:
            - available
        hubspot:
          type: object
          properties:
            available:
              type: boolean
            contacts:
              type: object
              properties:
                total:
                  type: integer
                newInPeriod:
                  type: integer
              required:
                - total
                - newInPeriod
            deals:
              type: object
              properties:
                total:
                  type: integer
                totalValue:
                  type: number
                byStage:
                  type: object
              required:
                - total
                - totalValue
                - byStage
          required:
            - available
      required:
        - property
        - dateRange
        - googleAds
        - metaAds
        - hubspot